Update docker-compose.yml
Updated all PUID and GUID to coordinate with the users that we made from setup.sh as well as the volumes to match the new folder structure. 
Important to note all the *arr services get /data mapped to a single volume rather than multiple as they only need specific folders in there. This preserves the ability to create hardlinks as advised by servarr wiki https://wiki.servarr.com/en/docker-guide